+++ daemon/src/ctrl_profiles/trait_impls.rs
@@ -2,6 +2,7 @@ use async_trait::async_trait;
 use log::warn;
 use rog_profiles::fan_curve_set::CurveData;
 use rog_profiles::fan_curve_set::FanCurveSet;
+use rog_profiles::FanCurveProfiles;
 use rog_profiles::Profile;
 use zbus::export::futures_util::lock::Mutex;
 use zbus::export::futures_util::StreamExt;
@@ -16,18 +17,12 @@ use crate::CtrlTask;
 
 use super::controller::CtrlPlatformProfile;
 
-static UNSUPPORTED_MSG: &str =
+const ZBUS_PATH: &str = "/org/asuslinux/Profile";
+const UNSUPPORTED_MSG: &str =
     "Fan curves are not supported on this laptop or you require a patched kernel";
 
-pub struct ProfileZbus {
-    inner: Arc<Mutex<CtrlPlatformProfile>>,
-}
-
-impl ProfileZbus {
-    pub fn new(inner: Arc<Mutex<CtrlPlatformProfile>>) -> Self {
-        Self { inner }
-    }
-}
+#[derive(Clone)]
+pub struct ProfileZbus(pub Arc<Mutex<CtrlPlatformProfile>>);
 
 #[dbus_interface(name = "org.asuslinux.Daemon")]
 impl ProfileZbus {
@@ -44,7 +39,7 @@ impl ProfileZbus {
     /// Toggle to next platform_profile. Names provided by `Profiles`.
     /// If fan-curves are supported will also activate a fan curve for profile.
     async fn next_profile(&mut self, #[zbus(signal_context)] ctxt: SignalContext<'_>) {
-        let mut ctrl = self.inner.lock().await;
+        let mut ctrl = self.0.lock().await;
         ctrl.set_next_profile()
             .unwrap_or_else(|err| warn!("{}", err));
         ctrl.save_config();
@@ -56,7 +51,7 @@ impl ProfileZbus {
 
     /// Fetch the active profile name
     async fn active_profile(&mut self) -> zbus::fdo::Result<Profile> {
-        let mut ctrl = self.inner.lock().await;
+        let mut ctrl = self.0.lock().await;
         ctrl.config.read();
         Ok(ctrl.config.active_profile)
     }
@@ -67,7 +62,7 @@ impl ProfileZbus {
         #[zbus(signal_context)] ctxt: SignalContext<'_>,
         profile: Profile,
     ) {
-        let mut ctrl = self.inner.lock().await;
+        let mut ctrl = self.0.lock().await;
         // Read first just incase the user has modified the config before calling this
         ctrl.config.read();
         Profile::set_profile(profile)
@@ -87,7 +82,7 @@ impl ProfileZbus {
 
     /// Get a list of profiles that have fan-curves enabled.
     async fn enabled_fan_profiles(&mut self) -> zbus::fdo::Result<Vec<Profile>> {
-        let mut ctrl = self.inner.lock().await;
+        let mut ctrl = self.0.lock().await;
         ctrl.config.read();
         if let Some(curves) = &ctrl.config.fan_curves {
             return Ok(curves.get_enabled_curve_profiles().to_vec());
@@ -102,7 +97,7 @@ impl ProfileZbus {
         profile: Profile,
         enabled: bool,
     ) -> zbus::fdo::Result<()> {
-        let mut ctrl = self.inner.lock().await;
+        let mut ctrl = self.0.lock().await;
         ctrl.config.read();
         return if let Some(curves) = &mut ctrl.config.fan_curves {
             curves.set_profile_curve_enabled(profile, enabled);
@@ -120,7 +115,7 @@ impl ProfileZbus {
 
     /// Get the fan-curve data for the currently active Profile
     async fn fan_curve_data(&mut self, profile: Profile) -> zbus::fdo::Result<FanCurveSet> {
-        let mut ctrl = self.inner.lock().await;
+        let mut ctrl = self.0.lock().await;
         ctrl.config.read();
         if let Some(curves) = &ctrl.config.fan_curves {
             let curve = curves.get_fan_curves_for(profile);
@@ -132,7 +127,7 @@ impl ProfileZbus {
     /// Set the fan curve for the specified profile.
     /// Will also activate the fan curve if the user is in the same mode.
     async fn set_fan_curve(&self, profile: Profile, curve: CurveData) -> zbus::fdo::Result<()> {
-        let mut ctrl = self.inner.lock().await;
+        let mut ctrl = self.0.lock().await;
         ctrl.config.read();
         if let Some(curves) = &mut ctrl.config.fan_curves {
             curves
@@ -154,7 +149,7 @@ impl ProfileZbus {
     /// Each platform_profile has a different default and the defualt can be read
     /// only for the currently active profile.
     async fn set_active_curve_to_defaults(&self) -> zbus::fdo::Result<()> {
-        let mut ctrl = self.inner.lock().await;
+        let mut ctrl = self.0.lock().await;
         ctrl.config.read();
         ctrl.set_active_curve_to_defaults()
             .map_err(|e| warn!("Profile::set_active_curve_to_defaults, {}", e))
@@ -168,7 +163,7 @@ impl ProfileZbus {
     /// Each platform_profile has a different default and the defualt can be read
     /// only for the currently active profile.
     async fn reset_profile_curves(&self, profile: Profile) -> zbus::fdo::Result<()> {
-        let mut ctrl = self.inner.lock().await;
+        let mut ctrl = self.0.lock().await;
         ctrl.config.read();
         let active = Profile::get_active_profile().unwrap_or(Profile::Balanced);
 
@@ -194,20 +189,19 @@ impl ProfileZbus {
 #[async_trait]
 impl crate::ZbusRun for ProfileZbus {
     async fn add_to_server(self, server: &mut Connection) {
-        Self::add_to_server_helper(self, "/org/asuslinux/Profile", server).await;
+        Self::add_to_server_helper(self, ZBUS_PATH, server).await;
     }
 }
 
 #[async_trait]
 impl CtrlTask for ProfileZbus {
+    fn zbus_path() -> &'static str {
+        ZBUS_PATH
+    }
+
     async fn create_tasks(&self, signal_ctxt: SignalContext<'static>) -> Result<(), RogError> {
-        let ctrl = self.inner.clone();
-        let mut watch = self
-            .inner
-            .lock()
-            .await
-            .platform
-            .monitor_platform_profile()?;
+        let ctrl = self.0.clone();
+        let mut watch = self.0.lock().await.platform.monitor_platform_profile()?;
 
         tokio::spawn(async move {
             let mut buffer = [0; 32];
@@ -230,6 +224,24 @@ impl CtrlTask for ProfileZbus {
                 .await;
         });
 
+        Ok(())
+    }
+}
+
+#[async_trait]
+impl crate::Reloadable for ProfileZbus {
+    /// Fetch the active profile and use that to set all related components up
+    async fn reload(&mut self) -> Result<(), RogError> {
+        let mut ctrl = self.0.lock().await;
+        let active = ctrl.config.active_profile;
+        if let Some(curves) = &mut ctrl.config.fan_curves {
+            if let Ok(mut device) = FanCurveProfiles::get_device() {
+                // There is a possibility that the curve was default zeroed, so this call initialises
+                // the data from system read and we need to save it after
+                curves.write_profile_curve_to_platform(active, &mut device)?;
+                ctrl.config.write();
+            }
+        }
         Ok(())
     }
 }